Change inheritance
The definition of members excludes now the underlying object.
The calculated result can be updated with the final result via one click.

When there is no final result the update creates a form for defining a
new row result. But this behaviour needs to be redefined in a further commit.

Implements #500